Moving toolbar into its own component
