+++ src/components/toolbar.sass
@@ -1,0 +1,44 @@
+@import vars
+
+.toolbar
+  position: absolute
+  top: $margin + $border-width + $workspace-titlebar-height
+  bottom: $workspace-statusbar-height
+  left: $margin - $workspace-toolbar-width - $border-width 
+  width: $workspace-toolbar-width
+
+  display: flex
+  flex-direction: column
+
+  .tool
+    width: $workspace-toolbar-width
+    height: $workspace-toolbar-width
+    background-color: #C9C9C9
+    display: flex
+    justify-content: center
+    align-items: center
+    text-align: center
+    font-size: 20px
+    font-weight: bold
+    margin-bottom: -$border-width
+    border: $border-width solid $border-color
+
+    &.select
+      border-top-left-radius: $border-radius
+
+    &.geometry
+      background-image: url("../assets/logo_tw.png")
+      background-size: contain
+      background-repeat: no-repeat
+
+    &.image
+      background-image: url("../assets/logo_image_tool.svg")
+      background-size: contain
+      background-repeat: no-repeat
+
+    &.delete
+      border-bottom-left-radius: $border-radius
+
+    &.active
+      color: #fff
+      background-color: #888888
